replace files from tab to the tictac new direction
