+++ app/(applications)/tictactoe/index.tsx
@@ -1,4 +1,4 @@
-import {useState} from "react"
+import React, {useState} from "react"
 import { StyleSheet, TouchableOpacity } from 'react-native';
 
 import EditScreenInfo from '@/components/EditScreenInfo';
