+++ app/(applications)/tictactoe/modal.tsx
@@ -1,15 +1,20 @@
-import { StyleSheet } from 'react-native';
+import React from 'react';
+import { StatusBar } from 'expo-status-bar';
+import { Platform, StyleSheet } from 'react-native';
 
 import EditScreenInfo from '@/components/EditScreenInfo';
 import { View } from '@/components/Themed';
 import { MonoText } from '@/components/StyledText';
 
-export default function TabTwoScreen() {
+export default function ModalScreen() {
   return (
     <View style={styles.container}>
-      <MonoText style={styles.title}>Tab Two</MonoText>
+      <MonoText style={styles.title}>Modal</MonoText>
       <View style={styles.separator} lightColor="#eee" darkColor="rgba(255,255,255,0.1)" />
-      <EditScreenInfo path="app/(tabs)/two.tsx" />
+      <EditScreenInfo path="app/modal.tsx" />
+
+      {/* Use a light status bar on iOS to account for the black space above the modal */}
+      <StatusBar style={Platform.OS === 'ios' ? 'light' : 'auto'} />
     </View>
   );
 }
