+++ app/(applications)/tictactoe/two.tsx
@@ -1,19 +1,16 @@
-import { StatusBar } from 'expo-status-bar';
-import { Platform, StyleSheet } from 'react-native';
+import React from 'react';
+import { StyleSheet } from 'react-native';
 
 import EditScreenInfo from '@/components/EditScreenInfo';
 import { View } from '@/components/Themed';
 import { MonoText } from '@/components/StyledText';
 
-export default function ModalScreen() {
+export default function TabTwoScreen() {
   return (
     <View style={styles.container}>
-      <MonoText style={styles.title}>Modal</MonoText>
+      <MonoText style={styles.title}>Tab Two</MonoText>
       <View style={styles.separator} lightColor="#eee" darkColor="rgba(255,255,255,0.1)" />
-      <EditScreenInfo path="app/modal.tsx" />
-
-      {/* Use a light status bar on iOS to account for the black space above the modal */}
-      <StatusBar style={Platform.OS === 'ios' ? 'light' : 'auto'} />
+      <EditScreenInfo path="app/(tabs)/two.tsx" />
     </View>
   );
 }
